moved transaction holder out into public class

+++ src/Rebus.Tests/Persistence/SqlServerFixtureBase.cs
@@ -1,7 +1,9 @@
+using System;
 using System.Collections.Generic;
 using System.Data.SqlClient;
 using NUnit.Framework;
 using log4net.Config;
+using System.Linq;
 
 namespace Rebus.Tests.Persistence
 {
@@ -40,6 +42,9 @@ namespace Rebus.Tests.Persistence
 
         protected void DeleteRows(string tableName)
         {
+            if (!GetTableNames()
+                     .Contains(tableName, StringComparer.InvariantCultureIgnoreCase)) return;
+
             ExecuteCommand("delete from " + tableName);
         }
